fix(restore): #2445 backup card shown after restore
